`await void` is "voiding" the await...
This seems like a workaround to stop typescript from complaining?
await void returns immediately and the responses are therefore not properly awaited causing some assertions to be skipped completely

Changed it to use proper (awaitable) promises and trying to simplify the nested ones to be on one code level using async / await.

Signed-off-by: Jannik Hollenbach <jannik@hollenbach.de>